Suppress more errors from systemd

+++ scripts/bin/toshy-systemd-remove.sh
@@ -21,29 +21,35 @@ echo -e "\nRemoving Toshy systemd services...\n"
 
 # systemctl "disable" automatically deletes all symlinks (might be bug)
 
-if systemctl --user --quiet is-active toshy-session-monitor.service; then
-    /usr/bin/systemctl --user stop toshy-session-monitor.service
-fi
+if [ -f "$HOME/.config/systemd/user/toshy-config.service" ]; then
+    if systemctl --user --quiet is-active toshy-session-monitor.service; then
+        /usr/bin/systemctl --user stop toshy-session-monitor.service
+    fi
 
-if systemctl --user --quiet is-enabled toshy-session-monitor.service; then
-    /usr/bin/systemctl --user disable toshy-session-monitor.service
-fi
-
-sleep $DELAY
-
-if systemctl --user --quiet is-active toshy-config.service; then
-    /usr/bin/systemctl --user stop toshy-config.service
-fi
-
-if systemctl --user --quiet is-enabled toshy-config.service; then
-    /usr/bin/systemctl --user disable toshy-config.service
+    if systemctl --user --quiet is-enabled toshy-session-monitor.service; then
+        /usr/bin/systemctl --user disable toshy-session-monitor.service
+    fi
+    rm -f "$HOME/.config/systemd/user/toshy-config.service"
 fi
 
 sleep $DELAY
 
-rm -f "$HOME/.config/systemd/user/toshy-config.service"
-rm -f "$HOME/.config/systemd/user/toshy-session-monitor.service"
-rm -f "$HOME/.config/autostart/Toshy_Import_Vars.desktop"
+if [ -f "$HOME/.config/systemd/user/toshy-session-monitor.service" ]; then
+    if systemctl --user --quiet is-active toshy-config.service; then
+        /usr/bin/systemctl --user stop toshy-config.service
+    fi
+
+    if systemctl --user --quiet is-enabled toshy-config.service; then
+        /usr/bin/systemctl --user disable toshy-config.service
+    fi
+    rm -f "$HOME/.config/systemd/user/toshy-session-monitor.service"
+fi
+
+sleep $DELAY
+
+if [ -f "$HOME/.config/autostart/Toshy_Import_Vars.desktop" ]; then
+    rm -f "$HOME/.config/autostart/Toshy_Import_Vars.desktop"
+fi
 
 sleep $DELAY
